+++ apps/pie-microsite/src/_11ty/shortcodes/tokensTable/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-trailing-spaces */
-const pieDesignTokens = require('@justeat/pie-design-tokens/dist/tokens.json');
+const normalizedPieDesignTokens = require('../../../_data/normalizeTokens');
 const pieTokenCategories = require('../../../tokenCategories.json');
 const { stringHelpers, objectHelpers, numberHelpers } = require('../../../utilities/helpers');
 const tokenTypes = require('../../../_data/tokenTypes');
@@ -37,6 +37,28 @@ const createTokenDisplayName = (tokenKey, tokenType) => {
     return shouldShowPrefix
         ? `${stringHelpers.capitaliseFirstLetter(tokenType)} ${capitalisedNameSegments.join(' ')}`
         : capitalisedNameSegments.join(' ');
+};
+
+/**
+ * Splits a font/typography token into a parsable css value
+ * @param {object} token - the token value i.e. {"size": "48|56","weight": "ExtraBold"}
+ * @returns {object} an object containing the font styles of the token.
+ */
+const splitFontAliasToken = token => {
+    const [fontSize, lineHeight] = token.size.split('|');
+
+    const fontWeightMap = {
+        Regular: 400,
+        Bold: 700,
+        ExtraBold: 800
+    };
+    
+    return {
+        fontSize, 
+        lineHeight,
+        fontWeight: fontWeightMap[token.weight],
+        textDecoration: token['text-decoration']
+    };
 };
 
 /**
@@ -105,6 +127,24 @@ const buildSpacingExample = token => {
 };
 
 /**
+* Builds an example font/typography element to show on the token list item
+ * @param {object} token - the token value i.e. {"size": "48|56","weight": "ExtraBold"}
+* @returns {string} - the typography example HTML string
+*/
+const buildFontExample = token => { 
+    const {
+        fontSize, lineHeight, fontWeight, textDecoration 
+    } = splitFontAliasToken(token);
+    const cssVariables = [
+    `--example-font-size: ${fontSize}px`,
+    `--example-font-line-height: ${lineHeight}px`,
+    `--example-font-weight: ${fontWeight}`,
+    textDecoration && `--example-font-text-decoration: ${textDecoration}`
+    ].filter(Boolean);
+
+    return `<div class="c-tokensTable-example--font" style="${cssVariables.join('; ')}">String</div>`;
+};
+/**
  * Builds an example element to display in the token list item.
  * This could be a color swatch, a representation of border radius or spacing etc.
  * @param {string} token - the token value i.e. #000, #ffffff, #000|0.85 or #000000|0.85
@@ -114,8 +154,9 @@ const buildSpacingExample = token => {
 const buildTokenExampleElement = (token, tokenType) => {
     const tokenExampleElementHandler = {
         [tokenTypes.COLOR]: buildColorExample,
-        [tokenTypes.SPACING]: buildSpacingExample,
-        [tokenTypes.RADIUS]: buildRadiusExample
+        [tokenTypes.FONT]: buildFontExample,
+        [tokenTypes.RADIUS]: buildRadiusExample,
+        [tokenTypes.SPACING]: buildSpacingExample
     };
 
     if (!tokenExampleElementHandler[tokenType]) {
@@ -199,7 +240,7 @@ const buildTokensListForCategory = (tokens, path, category, tokenType) => {
     const tokenListElements = tokensForCategory.map(key => buildTokenListElements({
         token: tokens[key],
         tokenScssName: createScssTokenName(key, tokenType),
-        tokenDisplayName: createTokenDisplayName(key, tokenType),
+        tokenDisplayName: tokenTypeMetadata[key].displayName ?? createTokenDisplayName(key, tokenType),
         tokenType,
         tokenMetadata: tokenTypeMetadata[key]
     }));
@@ -304,7 +345,7 @@ const buildCategoryListsWithParents = ({
  */
 const buildTokenLists = (path, tokenType) => {
     const isGlobal = path.includes('global');
-    const tokens = objectHelpers.getObjectPropertyByPath(pieDesignTokens, `theme.jet.${path}`);
+    const tokens = objectHelpers.getObjectPropertyByPath(normalizedPieDesignTokens, `theme.jet.${path}`);
     const parentCategories = getParentCategoriesForTokenType(`${tokenType}.${isGlobal ? 'global' : 'alias'}.parentCategories`);
     const regularCategories = objectHelpers.getObjectPropertyByPath(pieTokenCategories, path);